pyproject.toml
[project]
name = "inventory-tracker"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "django",
    "django-rest-framework>=0.1.0",
    "djangorestframework-simplejwt>=5.5.1",
    "mypy>=1.11",
    "python-dotenv>=1.1.1",
]
